FIT1045_PROJECT/HD_PROJECT/trial3/requirements.txt
streamlit
ollama
psutil
plotly
requests
PyPDF2
Pillow